AI Plant Guardian/requirements.txt
streamlit>=1.28.0
tensorflow>=2.13.0
opencv-python>=4.8.0
Pillow>=10.0.0
numpy>=1.24.0
matplotlib>=3.7.0
plotly>=5.15.0
pandas>=2.0.0
